+++ src/main/java/seedu/address/logic/parser/AddCommandParser.java
@@ -12,11 +12,8 @@ import java.util.stream.Stream;
 
 import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.Address;
-import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
-import seedu.address.model.person.Person;
-import seedu.address.model.person.Phone;
+import seedu.address.model.task.*;
+import seedu.address.model.task.Task;
 import seedu.address.model.tag.Tag;
 
 /**
@@ -44,9 +41,9 @@ public class AddCommandParser implements Parser<AddCommand> {
         Address address = ParserUtil.parseAddress(argMultimap.getValue(PREFIX_ADDRESS).get());
         Set<Tag> tagList = ParserUtil.parseTags(argMultimap.getAllValues(PREFIX_TAG));
 
-        Person person = new Person(name, phone, email, address, tagList);
+        Task task = new Task(name, phone, email, address, tagList);
 
-        return new AddCommand(person);
+        return new AddCommand(task);
     }
 
     /**
