+++ src/main/java/seedu/address/logic/parser/SchedulePlannerParser.java
@@ -24,7 +24,7 @@ import seedu.address.logic.parser.exceptions.ParseException;
 /**
  * Parses user input.
  */
-public class AddressBookParser {
+public class SchedulePlannerParser {
 
     /**
      * Used for initial separation of command word and args.
